SearchDialog 불필요한 리렌더링 최적화
- 컴포넌트들 분리하고memo 로 묶기
