+++ src/searchDialog/SelectDay.tsx
@@ -1,0 +1,33 @@
+import { memo } from "react";
+import {
+  FormControl,
+  FormLabel,
+  CheckboxGroup,
+  HStack,
+  Checkbox,
+} from "@chakra-ui/react";
+import { DAY_LABELS } from "../constants";
+
+interface Props {
+  days: string[];
+  handleDaysChange: (value: (string | number)[]) => void;
+}
+
+const SelectDay = ({ days, handleDaysChange }: Props) => {
+  return (
+    <FormControl>
+      <FormLabel>요일</FormLabel>
+      <CheckboxGroup value={days} onChange={handleDaysChange}>
+        <HStack spacing={4}>
+          {DAY_LABELS.map((day) => (
+            <Checkbox key={day} value={day}>
+              {day}
+            </Checkbox>
+          ))}
+        </HStack>
+      </CheckboxGroup>
+    </FormControl>
+  );
+};
+
+export default memo(SelectDay);
